Organize DTOs into subpackages
