+++ src/main/java/se/hydroleaf/controller/StatusController.java
@@ -4,9 +4,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import se.hydroleaf.dto.StatusAllAverageResponse;
-import se.hydroleaf.dto.StatusAverageResponse;
-import se.hydroleaf.dto.LiveNowSnapshot;
+import se.hydroleaf.dto.summary.StatusAllAverageResponse;
+import se.hydroleaf.dto.summary.StatusAverageResponse;
+import se.hydroleaf.dto.snapshot.LiveNowSnapshot;
 import se.hydroleaf.service.StatusService;
 
 @RestController
