+++ src/main/java/se/hydroleaf/dto/history/AggregatedHistoryResponse.java
@@ -1,4 +1,4 @@
-package se.hydroleaf.dto;
+package se.hydroleaf.dto.history;
 
 import java.time.Instant;
 import java.util.List;
