+++ src/main/java/se/hydroleaf/dto/summary/ActuatorStatusSummary.java
@@ -1,4 +1,4 @@
-package se.hydroleaf.dto;
+package se.hydroleaf.dto.summary;
 
 public record ActuatorStatusSummary(
         StatusAverageResponse airPump
