+++ src/main/java/se/hydroleaf/dto/summary/GrowSensorSummary.java
@@ -1,4 +1,4 @@
-package se.hydroleaf.dto;
+package se.hydroleaf.dto.summary;
 
 public record GrowSensorSummary(
         StatusAverageResponse light,
