+++ src/main/java/se/hydroleaf/scheduler/LiveFeedScheduler.java
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
-import se.hydroleaf.dto.LiveNowSnapshot;
+import se.hydroleaf.dto.snapshot.LiveNowSnapshot;
 import se.hydroleaf.service.StatusService;
 
 import java.time.Duration;
