+++ src/main/java/se/hydroleaf/service/RecordService.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import se.hydroleaf.dto.AggregatedHistoryResponse;
-import se.hydroleaf.dto.AggregatedSensorData;
-import se.hydroleaf.dto.TimestampValue;
+import se.hydroleaf.dto.history.AggregatedHistoryResponse;
+import se.hydroleaf.dto.history.AggregatedSensorData;
+import se.hydroleaf.dto.history.TimestampValue;
 import se.hydroleaf.model.ActuatorStatus;
 import se.hydroleaf.model.Device;
 import se.hydroleaf.model.LatestActuatorStatus;
