+++ src/main/java/se/hydroleaf/service/StatusService.java
@@ -3,13 +3,13 @@ package se.hydroleaf.service;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import se.hydroleaf.dto.ActuatorStatusSummary;
-import se.hydroleaf.dto.LiveNowSnapshot;
-import se.hydroleaf.dto.GrowSensorSummary;
-import se.hydroleaf.dto.SystemSnapshot;
-import se.hydroleaf.dto.WaterTankSummary;
-import se.hydroleaf.dto.StatusAllAverageResponse;
-import se.hydroleaf.dto.StatusAverageResponse;
+import se.hydroleaf.dto.summary.ActuatorStatusSummary;
+import se.hydroleaf.dto.snapshot.LiveNowSnapshot;
+import se.hydroleaf.dto.summary.GrowSensorSummary;
+import se.hydroleaf.dto.snapshot.SystemSnapshot;
+import se.hydroleaf.dto.summary.WaterTankSummary;
+import se.hydroleaf.dto.summary.StatusAllAverageResponse;
+import se.hydroleaf.dto.summary.StatusAverageResponse;
 import se.hydroleaf.model.Device;
 import se.hydroleaf.repository.AverageResult;
 import se.hydroleaf.repository.DeviceRepository;
