+++ src/test/java/se/hydroleaf/service/StatusServiceTest.java
@@ -6,10 +6,10 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.mockito.Spy;
-import se.hydroleaf.dto.LiveNowSnapshot;
-import se.hydroleaf.dto.StatusAllAverageResponse;
-import se.hydroleaf.dto.StatusAverageResponse;
-import se.hydroleaf.dto.SystemSnapshot;
+import se.hydroleaf.dto.snapshot.LiveNowSnapshot;
+import se.hydroleaf.dto.summary.StatusAllAverageResponse;
+import se.hydroleaf.dto.summary.StatusAverageResponse;
+import se.hydroleaf.dto.snapshot.SystemSnapshot;
 import se.hydroleaf.model.Device;
 import se.hydroleaf.repository.AverageResult;
 import se.hydroleaf.repository.DeviceRepository;
